final fixes to unit tests
